new method of computing DTW correlation. See description
Instead of using the DTW distance, we instead apply the warping to the query (predicted) timeseries, then take the R^2 (or hamming distance) of the warped timeseries. We do this because we don't care so much about *how* much we have to warp our timeseries, so long as we don't have to do it too much. We just want to know how well they could line up if they were lined up. Also added regularization to the dtw fit in the form of jittering the points slightly.